FEAT: refactoring PyFibreGUI
